Some fixes just to compile
This commit includes some minor fixes so as to make

In this sense, I have concerns about two fields of the
tQucsSettings structure I've added ...just because somewhere the
compiler was complaining about the following fields:

QDir OctaveBinDir;   // dir of octave executable
QString OctaveExecutable;//Are these three fields just the same?
QString QucsOctave;

What is the purpose of these fields? Are all of them pointing to
the Octave executable?